# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgenticIoT Capability API
  version: 0.1.0
  description: >-
    Stable northbound API for discovering, observing, and invoking physical-device
    capabilities. Management and Edge APIs will be specified separately.
servers:
  - url: /v1
security:
  - bearerAuth: []
tags:
  - name: Things
  - name: Commands
  - name: Events
paths:
  /things:
    get:
      operationId: listThings
      tags: [Things]
      summary: Discover authorized things
      parameters:
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
        - name: domain_ref
          in: query
          schema: {type: string}
        - name: space_ref
          in: query
          schema: {type: string}
        - name: capability
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Authorized things
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThingList"
        default:
          $ref: "#/components/responses/Problem"
  /things/{thingId}:
    get:
      operationId: getThing
      tags: [Things]
      summary: Get a canonical Thing description
      parameters:
        - $ref: "#/components/parameters/ThingId"
      responses:
        "200":
          description: Thing description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thing"
        default:
          $ref: "#/components/responses/Problem"
  /things/{thingId}/state:
    get:
      operationId: getThingState
      tags: [Things]
      summary: Read the latest authorized state projection
      parameters:
        - $ref: "#/components/parameters/ThingId"
        - name: fresh
          in: query
          description: Ask the runtime to attempt a new physical read.
          schema: {type: boolean, default: false}
      responses:
        "200":
          description: State projection with observation metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThingState"
        default:
          $ref: "#/components/responses/Problem"
  /things/{thingId}/actions/{actionName}:
    post:
      operationId: invokeThingAction
      tags: [Things]
      summary: Request a durable asynchronous physical action
      parameters:
        - $ref: "#/components/parameters/ThingId"
        - name: actionName
          in: path
          required: true
          schema: {type: string}
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 200}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeActionRequest"
      responses:
        "202":
          description: Command durably accepted
          headers:
            Location:
              description: URL of the durable Command resource.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        default:
          $ref: "#/components/responses/Problem"
  /commands/{commandId}:
    get:
      operationId: getCommand
      tags: [Commands]
      summary: Get command status and available receipts
      parameters:
        - $ref: "#/components/parameters/CommandId"
      responses:
        "200":
          description: Durable Command resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        default:
          $ref: "#/components/responses/Problem"
  /commands/{commandId}/receipts:
    get:
      operationId: listCommandReceipts
      tags: [Commands]
      summary: List immutable execution receipts
      parameters:
        - $ref: "#/components/parameters/CommandId"
      responses:
        "200":
          description: Ordered receipts
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/ExecutionReceipt"}
        default:
          $ref: "#/components/responses/Problem"
  /events:
    get:
      operationId: subscribeEvents
      tags: [Events]
      summary: Subscribe to authorized events using Server-Sent Events
      parameters:
        - name: cursor
          in: query
          schema: {type: string}
        - name: thing_id
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Server-Sent Event stream with stable event IDs
          content:
            text/event-stream:
              schema: {type: string}
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ThingId:
      name: thingId
      in: path
      required: true
      schema: {type: string}
    CommandId:
      name: commandId
      in: path
      required: true
      schema: {type: string}
  responses:
    Problem:
      description: Problem Details response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ThingList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/ThingSummary"}
        next_cursor:
          type: [string, "null"]
    ThingSummary:
      type: object
      required: [id, model_id, lifecycle_status, reachability]
      properties:
        id: {type: string}
        model_id: {type: string}
        title: {type: string}
        domain_ref: {type: [string, "null"]}
        space_ref: {type: [string, "null"]}
        lifecycle_status:
          type: string
          enum: [commissioning, active, suspended, retiring, retired]
        reachability:
          type: string
          enum: [online, sleeping, offline, unknown]
    Thing:
      allOf:
        - $ref: "#/components/schemas/ThingSummary"
        - type: object
          required: [model_version, properties, actions, events]
          properties:
            model_version: {type: string}
            properties:
              type: object
              additionalProperties: {$ref: "#/components/schemas/PropertySpec"}
            actions:
              type: object
              additionalProperties: {$ref: "#/components/schemas/ActionSpec"}
            events:
              type: object
              additionalProperties: {$ref: "#/components/schemas/EventSpec"}
    PropertySpec:
      type: object
      required: [schema, readable]
      properties:
        schema: {type: object, additionalProperties: true}
        readable: {type: boolean}
        writable: {type: boolean, default: false}
        unit: {type: string}
    ActionSpec:
      type: object
      required: [input_schema, risk, confirmation]
      properties:
        input_schema: {type: object, additionalProperties: true}
        output_schema: {type: object, additionalProperties: true}
        risk: {type: string, enum: [low, medium, high, critical]}
        confirmation:
          type: string
          enum: [adapter_ack, observed_state, device_event, none]
        offline_policy:
          type: string
          enum: [local_always, cached_authorization, online_required, local_only]
          default: online_required
    EventSpec:
      type: object
      required: [data_schema]
      properties:
        data_schema: {type: object, additionalProperties: true}
    ThingState:
      type: object
      required: [thing_id, properties]
      properties:
        thing_id: {type: string}
        properties:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ObservedValue"
    ObservedValue:
      type: object
      required: [value, source_ref, observed_at, received_at, freshness]
      properties:
        value: {}
        source_ref: {type: string}
        observed_at: {type: string, format: date-time}
        received_at: {type: string, format: date-time}
        source_version: {type: [string, "null"]}
        freshness: {type: string, enum: [fresh, aging, stale, unknown]}
        quality: {type: string, enum: [good, uncertain, bad, unknown]}
    InvokeActionRequest:
      type: object
      required: [input]
      properties:
        input: {type: object, additionalProperties: true}
        deadline: {type: string, format: date-time}
        context:
          $ref: "#/components/schemas/RequestContext"
    RequestContext:
      type: object
      properties:
        subject_ref: {type: string}
        domain_ref: {type: string}
        space_ref: {type: string}
        agent_ref: {type: string}
        origin_ref: {type: string}
        authorization_ref: {type: string}
        expires_at: {type: string, format: date-time}
    Command:
      type: object
      required: [id, thing_id, action, status, created_at, trace_id, receipts]
      properties:
        id: {type: string}
        thing_id: {type: string}
        action: {type: string}
        status:
          type: string
          enum: [accepted, running, succeeded, failed, unknown, expired]
        created_at: {type: string, format: date-time}
        deadline: {type: [string, "null"], format: date-time}
        trace_id: {type: string}
        receipts:
          type: array
          items: {$ref: "#/components/schemas/ExecutionReceipt"}
    ExecutionReceipt:
      type: object
      required: [command_id, sequence, stage, occurred_at, trace_id]
      properties:
        command_id: {type: string}
        sequence: {type: integer, minimum: 1}
        stage:
          type: string
          enum: [received, rejected, accepted, dispatched, acknowledged, confirmed, failed, timed_out, expired, diverged]
        executor_ref: {type: [string, "null"]}
        occurred_at: {type: string, format: date-time}
        evidence: {type: object, additionalProperties: true}
        error_code: {type: [string, "null"]}
        trace_id: {type: string}
    Problem:
      type: object
      required: [type, title, status, code, trace_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 100, maximum: 599}
        detail: {type: string}
        code: {type: string}
        trace_id: {type: string}